Started AWS integration. Finished S3 integration for resume uploading. Now working on bedrock agent for extracting job information from URLs!

+++ src/main/java/com/kenneth/nextrole/Service/ResumeService.java
@@ -4,28 +4,34 @@ package com.kenneth.nextrole.Service;
 import com.kenneth.nextrole.Model.Resume;
 import com.kenneth.nextrole.Model.User;
 import com.kenneth.nextrole.Repository.ResumeRepository;
+import com.kenneth.nextrole.awsApps.S3Service;
 import com.kenneth.nextrole.dto.resume.CreateResumeRequest;
 import com.kenneth.nextrole.dto.resume.ResumeResponse;
 import com.kenneth.nextrole.dto.resume.UpdateResumeRequest;
+import com.kenneth.nextrole.dto.resume.ViewSingleResumeResponse;
 import com.kenneth.nextrole.exception.ResumeNotFoundException;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import org.apache.commons.io.FilenameUtils;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-
+import java.util.UUID;
 @Service
 public class ResumeService {
 
     private final ResumeRepository resumeRepository;
+    private final S3Service service;
 
-    public ResumeService (ResumeRepository resumeRepository){
+
+    public ResumeService (ResumeRepository resumeRepository, S3Service service){
         this.resumeRepository = resumeRepository;
+        this.service = service;
     }
-
 
     private ResumeResponse toResponse(Resume resume) {
         return ResumeResponse.builder()
@@ -36,8 +42,17 @@ public class ResumeService {
                 .build();
     }
 
+    private ViewSingleResumeResponse singleResumeResponse(Resume resume, String url){
+        return ViewSingleResumeResponse.builder()
+                .id(resume.getId())
+                .resumeTitle(resume.getResumeTitle())
+                .fileSize(resume.getFileSize())
+                .uploadedAt(resume.getUploadedAt())
+                .url(url)
+                .build();
+    }
 
-    public ResumeResponse getResume(Long resumeId, User user) {
+    public ViewSingleResumeResponse getResume(Long resumeId, User user) {
 
         Resume resume = resumeRepository.findById(resumeId)
                 .orElseThrow(() -> new ResumeNotFoundException("Resume not found"));
@@ -46,17 +61,33 @@ public class ResumeService {
             throw new AccessDeniedException("Not authorized");
         }
 
-        return toResponse(resume);
+        String url = service.getPresignedUrl(resume.getS3ObjectKey());
+
+        return singleResumeResponse(resume, url);
     }
 
+    /*
+    Integrate the s3 functions here
+     */
     @Transactional
     public ResumeResponse createResume (CreateResumeRequest request, User user){
+
+        String objectKey = genObjectKey(user, request.getFile());
+
+        try {
+            service.uploadResume(request.getFile(), objectKey);
+        } catch (Exception e){
+            throw new RuntimeException("Could not upload file");
+        }
+
         Resume resume = Resume.builder().user(user).
-                fileSize(request.getFileSize()).
+                fileSize(request.getFile().getSize()).
                 resumeTitle(request.getResumeTitle()).
-                uploadedAt(LocalDateTime.now()).build();
+                uploadedAt(LocalDateTime.now()).s3ObjectKey(objectKey).
+                build();
 
         resume = resumeRepository.save(resume);
+
 
 
         return toResponse(resume);
@@ -67,6 +98,14 @@ public class ResumeService {
     Prevents having to make multiple api calls.
      */
 
+
+    //building out the object key to send back to S3.
+    private String genObjectKey(User user, MultipartFile file){
+        String uuid = UUID.randomUUID().toString();
+        String original_file = file.getOriginalFilename();
+        String extension = FilenameUtils.getExtension(original_file);
+        return "users/"+user.getId()+"resumes/"+uuid+"/"+extension;
+    }
     @Transactional
     public ResumeResponse updateResume(UpdateResumeRequest request, User user, Long resumeId){
         Resume resume = resumeRepository.findById(resumeId).
@@ -89,14 +128,20 @@ public class ResumeService {
        resume = resumeRepository.save(resume);
         return toResponse(resume);
     }
-
-
     /*
     may rewrite this idk
      */
 
     @Transactional
     public List<ResumeResponse> deleteUserResume(Long id, User user) {
+        Resume resume = resumeRepository.findById(id).
+                orElseThrow(() -> new ResumeNotFoundException("Resume not found"));
+
+        if(resume.getUser() != user){
+            throw new AccessDeniedException("This resume is not yours");
+        }
+        service.deleteResume(resume.getS3ObjectKey());
+
         resumeRepository.deleteByIdAndUserId(id, user.getId());
 
         return resumeRepository.findByUserId(user.getId())
